Retirar Telas splashScreem
