+++ lib/components/foundation_form/appbar_foundation.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:path/path.dart';
 
 class AppBarFoundation extends StatelessWidget {
   var icon;
@@ -8,6 +9,7 @@ class AppBarFoundation extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AppBar(
+      backgroundColor: Theme.of(context).primaryColor,
       automaticallyImplyLeading: false,
       leading: FlatButton(
         onPressed: () {
