+++ lib/components/foundation_form/body_foundation.dart
@@ -11,7 +11,7 @@ class BodyFoundation extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      color: Colors.blueAccent[200],
+      color: Theme.of(context).primaryColor,
       child: Column(
         children: <Widget>[
           Center(
