+++ lib/screens/reg_my_pet/my_pet_info_third.dart
@@ -1,20 +1,34 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:path/path.dart';
 import 'package:petsaojoao/components/foundation_form/appbar_foundation.dart';
 import 'package:petsaojoao/components/foundation_form/body_foundation.dart';
 import 'package:petsaojoao/components/foundation_form/button_confirmForm.dart';
 import 'package:petsaojoao/components/foundation_form/data_security_info.dart';
+import 'package:petsaojoao/models/back_reg_my_pet/camera_initializer.dart';
 import 'package:petsaojoao/models/validators/castration_validator.dart';
 import 'package:petsaojoao/models/validators/microchip_validator.dart';
 import 'package:petsaojoao/models/validators/size_pet_validator.dart';
+import 'package:petsaojoao/screens/dashboard/dashboard.dart';
+import 'package:petsaojoao/services/repo_reg_my_pet/api_rest_reg_my_pet.dart';
 
-import 'end_form_reg_my_pet.dart';
+import 'my_pet_info_first.dart';
+import 'my_pet_info_second.dart';
 
 TextEditingController _sizePetController = new TextEditingController();
 TextEditingController _castrationController = new TextEditingController();
 TextEditingController _microchipController = new TextEditingController();
 var _idSizePet;
 var _statusCastration;
+
+var name = MyPetInfoFirst().getName();
+var breedId = MyPetInfoSecond().getIdBreeds();
+var genere = MyPetInfoFirst().getIdGender();
+var dateBirth = MyPetInfoSecond().getBirthDatePet();
+var coatColorId = MyPetInfoSecond().getCoatColors();
+var size = MyPetInfoThird().getSizePet();
+var castrated = MyPetInfoThird().getStatusCastration();
+var microchipSerial = MyPetInfoThird().getMicrochip();
 
 class MyPetInfoThird extends StatefulWidget {
   String getSizePet() {
@@ -30,7 +44,11 @@ class MyPetInfoThird extends StatefulWidget {
   }
 
   String getMicrochip() {
-    return _microchipController.text;
+    if (_microchipController.text == '') {
+      return null;
+    } else {
+      return _microchipController.text;
+    }
   }
 
   @override
@@ -104,14 +122,29 @@ class _FormMyPetInfoThirdState extends State<FormMyPetInfoThird> {
           ),
           ButtonConfirmForm(() {
             if (_myPetInfoThirdFormKey.currentState.validate()) {
-              Navigator.push(context,
-                  MaterialPageRoute(builder: (context) => EndFormRegMypet()));
+              sendData(context);
+              // Navigator.push(context,
+              //     MaterialPageRoute(builder: (context) => EndFormRegMypet()));
               //CameraInitializer().init(context);
             }
           })
         ]),
       ),
     );
+  }
+}
+
+sendData(context) async {
+  DrawerBottonSendData().show(context);
+  var result = await ApiRestRegMyPet.post(name, breedId, genere,
+      dateBirth.toString(), coatColorId, size, castrated, microchipSerial);
+
+  if (result == null) {
+    Navigator.pop(context);
+    DrawerBottonError().show(context);
+  } else {
+    Navigator.pop(context);
+    DrawerBottonPositive().show(context);
   }
 }
 
@@ -298,5 +331,162 @@ class _DropDownCastrationState extends State<DropDownCastration> {
         border: OutlineInputBorder(),
       ),
     );
+  }
+}
+
+class DrawerBottonSendData {
+  var _labelSend = "Enviando Dados";
+  void show(context) {
+    showModalBottomSheet<void>(
+        context: context,
+        builder: (BuildContext context) {
+          return Container(
+              color: Colors.blueAccent[200],
+              child: Padding(
+                  padding: const EdgeInsets.all(32.0),
+                  child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: <Widget>[
+                              CircularProgressIndicator(
+                                valueColor: new AlwaysStoppedAnimation<Color>(
+                                    Colors.white),
+                              ),
+                              Text(_labelSend,
+                                  textAlign: TextAlign.center,
+                                  style: TextStyle(
+                                      color: Colors.white, fontSize: 24.0)),
+                            ])
+                      ])));
+        });
+  }
+}
+
+class DrawerBottonError {
+  final _labelError = "Não foi possível enviar seus dados";
+  final buttonTryAgain = "Tentar Novamente";
+  final buttonCancel = "Cancelar";
+
+  void show(context) {
+    showModalBottomSheet<void>(
+        context: context,
+        builder: (BuildContext context) {
+          return Container(
+              color: Colors.redAccent,
+              child: Padding(
+                  padding: const EdgeInsets.all(32.0),
+                  child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Icon(
+                          FontAwesomeIcons.timesCircle,
+                          color: Colors.white,
+                          size: 40,
+                        ),
+                        SizedBox(
+                          height: 10,
+                        ),
+                        Text(_labelError,
+                            textAlign: TextAlign.center,
+                            style:
+                                TextStyle(color: Colors.white, fontSize: 24.0)),
+                        SizedBox(
+                          height: 20,
+                        ),
+                        Row(
+                            mainAxisAlignment: MainAxisAlignment.end,
+                            children: <Widget>[
+                              FlatButton(
+                                color: Colors.redAccent,
+                                textColor: Colors.white,
+                                padding: EdgeInsets.all(8.0),
+                                onPressed: () async {
+                                  await sendData(context);
+                                },
+                                child: Text(
+                                  buttonTryAgain.toUpperCase(),
+                                  style: TextStyle(
+                                      fontSize: 10.0,
+                                      fontWeight: FontWeight.bold),
+                                ),
+                              ),
+                              FlatButton(
+                                color: Colors.redAccent,
+                                textColor: Colors.brown[200],
+                                padding: EdgeInsets.all(8.0),
+                                onPressed: () {
+                                  Navigator.pushReplacement(
+                                      context,
+                                      MaterialPageRoute(
+                                          builder: (context) => Dashboard()));
+                                },
+                                child: Text(
+                                  buttonCancel.toUpperCase(),
+                                  style: TextStyle(
+                                      fontSize: 10.0,
+                                      fontWeight: FontWeight.bold),
+                                ),
+                              ),
+                            ])
+                      ])));
+        });
+  }
+}
+
+class DrawerBottonPositive {
+  final _labelPositive = "Agora vamos precisar de umas fotos do seu pet";
+  final buttonLater = "OK";
+
+  void show(context) {
+    showModalBottomSheet<void>(
+        context: context,
+        builder: (BuildContext context) {
+          return Container(
+              color: Colors.green,
+              child: Padding(
+                  padding: const EdgeInsets.all(32.0),
+                  child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Icon(
+                          FontAwesomeIcons.checkCircle,
+                          color: Colors.white,
+                          size: 40,
+                        ),
+                        SizedBox(
+                          height: 10,
+                        ),
+                        Text(_labelPositive,
+                            textAlign: TextAlign.center,
+                            style:
+                                TextStyle(color: Colors.white, fontSize: 24.0)),
+                        SizedBox(
+                          height: 20,
+                        ),
+                        Row(
+                            mainAxisAlignment: MainAxisAlignment.end,
+                            children: <Widget>[
+                              FlatButton(
+                                color: Colors.green,
+                                textColor: Colors.white,
+                                padding: EdgeInsets.all(8.0),
+                                onPressed: () {
+                                  CameraInitializer().init(context);
+                                },
+                                child: Text(
+                                  buttonLater.toUpperCase(),
+                                  style: TextStyle(
+                                      fontSize: 15.0,
+                                      fontWeight: FontWeight.bold),
+                                ),
+                              ),
+                            ])
+                      ])));
+        });
   }
 }
