+++ lib/screens/register_tutor/personal_info.dart
@@ -1,5 +1,8 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_masked_text/flutter_masked_text.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:petsaojoao/components/foundation_form/appbar_foundation.dart';
 import 'package:petsaojoao/components/foundation_form/body_foundation.dart';
 import 'package:petsaojoao/components/foundation_form/button_confirmForm.dart';
@@ -7,8 +10,10 @@ import 'package:petsaojoao/components/foundation_form/data_security_info.dart';
 import 'package:petsaojoao/models/validators/cpf_validator.dart';
 import 'package:petsaojoao/models/validators/full_name_validator.dart';
 import 'package:petsaojoao/models/validators/rg_validator.dart';
+import 'package:petsaojoao/services/repo_reg_tutor/api_rest_tutor.dart';
 
 import 'andress_info.dart';
+import 'contact_info.dart';
 
 TextEditingController _nameController = new TextEditingController();
 var _rgController = new MaskedTextController(mask: '00.000.000');
@@ -23,6 +28,22 @@ class _PersonalInfoState extends State<PersonalInfo> {
   String steps = "1 de 3";
   String questionTittle = "Quem é você?";
   var exitIcon = Icons.clear;
+
+  void showDrawer() {
+    DrawerBottonPositive().show(context);
+  }
+
+  startCadTutorTimer() async {
+    var _duration = new Duration(milliseconds: 500);
+    return new Timer(_duration, showDrawer);
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    startCadTutorTimer();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -96,10 +117,13 @@ class _FormPersonalState extends State<FormPersonal> {
 
   FocusNode focusNameForRg;
   FocusNode foucusRgForCpf;
+  sendData() async {
+    //DrawerBottonPositive().show(context);
+  }
 
   void initState() {
     super.initState();
-
+    //sendData();
     focusNameForRg = FocusNode();
     foucusRgForCpf = FocusNode();
   }
@@ -170,5 +194,60 @@ class _FormPersonalState extends State<FormPersonal> {
             ],
           )),
     );
+  }
+}
+
+class DrawerBottonPositive {
+  final _labelPositive = "Precisamos dos seus dados";
+  final buttonYes = "Sim";
+  final buttonOK = "OK";
+
+  void show(context) {
+    showModalBottomSheet<void>(
+        context: context,
+        builder: (BuildContext context) {
+          return Container(
+              color: Theme.of(context).primaryColor,
+              child: Padding(
+                  padding: const EdgeInsets.all(32.0),
+                  child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Icon(
+                          Icons.info,
+                          color: Colors.white,
+                          size: 40,
+                        ),
+                        SizedBox(
+                          height: 10,
+                        ),
+                        Text(_labelPositive,
+                            textAlign: TextAlign.center,
+                            style:
+                                TextStyle(color: Colors.white, fontSize: 24.0)),
+                        SizedBox(
+                          height: 20,
+                        ),
+                        Row(
+                            mainAxisAlignment: MainAxisAlignment.end,
+                            children: <Widget>[
+                              FlatButton(
+                                color: Colors.blue,
+                                textColor: Colors.white,
+                                padding: EdgeInsets.all(8.0),
+                                onPressed: () {
+                                  Navigator.pop(context);
+                                },
+                                child: Text(
+                                  buttonOK.toUpperCase(),
+                                  style: TextStyle(
+                                      fontSize: 15.0,
+                                      fontWeight: FontWeight.bold),
+                                ),
+                              ),
+                            ])
+                      ])));
+        });
   }
 }
