+++ lib/screens/register_tutor/register_tutor.dart
@@ -1,6 +1,7 @@
 //Acompanhe desing do projeto aqui --> https://www.figma.com/file/GYFrt79mzIbOUXXmFyDgwL/Material-Baseline-Design-Kit?node-id=38%3A5814
 import 'dart:async';
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:petsaojoao/components/foundation_form/splash_screen_foundation.dart';
 import 'package:petsaojoao/screens/register_tutor/personal_info.dart';
 
@@ -11,8 +12,9 @@ class RegisterTutor extends StatefulWidget {
 
 class _RegisterTutorState extends State<RegisterTutor> {
   void navigationToNextPage() {
-    //Navigator.push(context, _createRoute());
-    Navigator.pushReplacement(context, _createRoute());
+    // DrawerBottonPositive().show(context);
+    Navigator.push(context, _createRoute());
+    //Navigator.pushReplacement(context, _createRoute());
   }
 
   startCadTutorTimer() async {
@@ -55,4 +57,59 @@ Route _createRoute() {
       );
     },
   );
+}
+
+class DrawerBottonPositive {
+  final _labelPositive = "Deseja cadastrar seu pet agora?";
+  final buttonYes = "Sim";
+  final buttonLater = "Mais tarde";
+
+  void show(context) {
+    showModalBottomSheet<void>(
+        context: context,
+        builder: (BuildContext context) {
+          return Container(
+              color: Colors.green,
+              child: Padding(
+                  padding: const EdgeInsets.all(32.0),
+                  child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Icon(
+                          FontAwesomeIcons.checkCircle,
+                          color: Colors.white,
+                          size: 40,
+                        ),
+                        SizedBox(
+                          height: 10,
+                        ),
+                        Text(_labelPositive,
+                            textAlign: TextAlign.center,
+                            style:
+                                TextStyle(color: Colors.white, fontSize: 24.0)),
+                        SizedBox(
+                          height: 20,
+                        ),
+                        Row(
+                            mainAxisAlignment: MainAxisAlignment.end,
+                            children: <Widget>[
+                              FlatButton(
+                                color: Colors.green,
+                                textColor: Colors.brown[200],
+                                padding: EdgeInsets.all(8.0),
+                                onPressed: () {
+                                  Navigator.pop(context);
+                                },
+                                child: Text(
+                                  buttonLater.toUpperCase(),
+                                  style: TextStyle(
+                                      fontSize: 15.0,
+                                      fontWeight: FontWeight.bold),
+                                ),
+                              ),
+                            ])
+                      ])));
+        });
+  }
 }
