+++ lib/services/base_url.dart
@@ -1,1 +1,1 @@
-const BASE_URL = 'http://10.0.2.2:3000';
+const BASE_URL = 'http://192.168.0.114:3000';
